bugfix: renamed import and correctly specified path
