+++ src/utils/IsGeometry.ts
